transactions: add TxMatchesFilter2 function for gomobile support

+++ account_mixer.go
@@ -20,7 +20,7 @@ const (
 	ShuffleServer      = "cspp.decred.org"
 	MainnetShufflePort = "5760"
 	TestnetShufflePort = "15760"
-	MixedAccountBranch = udb.ExternalBranch
+	MixedAccountBranch = int32(udb.ExternalBranch)
 )
 
 func (mw *MultiWallet) AddAccountMixerNotificationListener(accountMixerNotificationListener AccountMixerNotificationListener, uniqueIdentifier string) error {
@@ -198,7 +198,7 @@ func (mw *MultiWallet) StartAccountMixer(walletID int, walletPassphrase string) 
 	}
 
 	tb.AccessConfig(func(c *ticketbuyer.Config) {
-		c.MixedAccountBranch = MixedAccountBranch
+		c.MixedAccountBranch = uint32(MixedAccountBranch)
 		c.MixedAccount = uint32(mixedAccount)
 		c.ChangeAccount = uint32(unmixedAccount)
 		c.CSPPServer = ShuffleServer + ":" + shufflePort
